#420 Move security feature implementation
