+++ neclib/core/__init__.py
@@ -12,6 +12,7 @@ make it hard to implement new features.
 from . import environ  # noqa: F401
 from . import formatting  # noqa: F401
 from . import logic  # noqa: F401
+from . import security  # noqa: F401
 from . import type_aliases  # noqa: F401
 from . import units  # noqa: F401
 from .configuration import *  # noqa: F401, F403
