+++ neclib/utils/__init__.py
@@ -4,5 +4,4 @@ from .array_utils import *  # noqa: F401, F403
 from .data_utils import *  # noqa: F401, F403
 from .math_utils import *  # noqa: F401, F403
 from .quantity_utils import *  # noqa: F401, F403
-from .security_utils import *  # noqa: F401, F403
 from .simulator_utils import *  # noqa: F401, F403
